Shift end date when start date is shifted on lesson plan event form
