+++ client/app/bundles/course/lesson-plan/containers/EventFormDialog/index.jsx
@@ -2,10 +2,11 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
-import { isPristine } from 'redux-form';
+import { getFormValues, isPristine } from 'redux-form';
 import FormDialogue from 'lib/components/FormDialogue';
 import { formNames } from 'course/lesson-plan/constants';
 import * as actionCreators from 'course/lesson-plan/actions';
+import * as libActionCreators from 'lib/actions';
 import EventForm from './EventForm';
 
 const EventFormDialog = ({
@@ -16,11 +17,10 @@ const EventFormDialog = ({
   onSubmit,
   pristine,
   dispatch,
+  formValues,
 }) => {
-  const {
-    hideEventForm,
-    submitEventForm,
-  } = bindActionCreators(actionCreators, dispatch);
+  const { hideEventForm, submitEventForm } = bindActionCreators(actionCreators, dispatch);
+  const { shiftEndDate } = bindActionCreators(libActionCreators, dispatch);
 
   return (
     <FormDialogue
@@ -31,7 +31,7 @@ const EventFormDialog = ({
       disabled={disabled}
       hideForm={hideEventForm}
     >
-      <EventForm {...{ initialValues, onSubmit, disabled }} />
+      <EventForm {...{ initialValues, onSubmit, disabled, shiftEndDate, formValues }} />
     </FormDialogue>
   );
 };
@@ -56,6 +56,7 @@ EventFormDialog.propTypes = {
     end_at: PropTypes.string,
     published: PropTypes.bool,
   }),
+  formValues: PropTypes.shape(),
   onSubmit: PropTypes.func.isRequired,
   pristine: PropTypes.bool.isRequired,
   dispatch: PropTypes.func.isRequired,
@@ -64,4 +65,5 @@ EventFormDialog.propTypes = {
 export default connect(({ eventForm, ...state }) => ({
   ...eventForm,
   pristine: isPristine(formNames.EVENT)(state),
+  formValues: getFormValues(formNames.EVENT)(state),
 }))(EventFormDialog);
