+++ client/app/lib/actions/index.js
@@ -1,4 +1,5 @@
 import actionTypes from 'lib/constants';
+import { change } from 'redux-form';
 
 export function setNotification(message) {
   return (dispatch) => {
@@ -20,5 +21,21 @@ export function showDeleteConfirmation(onConfirm) {
       dispatch(resetDeleteConfirmation());
     };
     dispatch({ type: actionTypes.SHOW_DELETE_CONFIRMATION, onConfirm: confirmAndDismiss });
+  };
+}
+
+export function shiftEndDate(formName, newStartAt, oldValues, startAtField = 'start_at', endAtField = 'end_at') {
+  return (dispatch) => {
+    const { [startAtField]: oldStartAt, [endAtField]: oldEndAt } = oldValues;
+    if (!oldStartAt || !oldEndAt) { return; }
+    const oldStartTime = oldStartAt.getTime();
+    const oldEndTime = oldEndAt.getTime();
+
+    // if start time is before end time, allow user to clear the error
+    if (oldStartTime <= oldEndTime) {
+      const newStartTime = newStartAt.getTime();
+      const newEndAt = new Date(oldEndTime + (newStartTime - oldStartTime));
+      dispatch(change(formName, endAtField, newEndAt));
+    }
   };
 }
